% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_functions.R
\docType{data}
\name{cawthron_palettes}
\alias{cawthron_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
cawthron_palettes
}
\description{
Official Cawthron colour palettes. Use \code{\link{get_pal}} to extract palettes.
}
\keyword{datasets}
