% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_functions.R
\name{get_pal}
\alias{get_pal}
\title{Select Cawthron palette}
\usage{
get_pal(name)
}
\arguments{
\item{name}{Name of desired palette. Choices are: \code{Hihi}, \code{Hoiho}, \code{Kaka}, \code{Kakariki}, and \code{Kakapo}.}
}
\value{
A vector of colours.
}
\description{
Return official Cawthron colour palettes.
}
\examples{
get_pal("Hihi")
}
