% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_functions.R
\name{get_pal}
\alias{get_pal}
\title{Select Cawthron palette}
\usage{
get_pal(name)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\itemize{
\item categorical: \code{caw_cat_1}, \code{caw_cat_2}, \code{caw_cat_3}
\item diverging: \code{caw_div_1}, \code{caw_div_2}, \code{caw_div_3}
\item sequential: \code{caw_seq_1}, \code{caw_seq_2}, \code{caw_seq_3}, \code{caw_seq_4}, \code{caw_seq_5}
}}
}
\value{
A vector of colours.
}
\description{
Return official Cawthron colour palettes.
}
\examples{
get_pal("caw_div_1")
}
