% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_functions.R
\name{print_pal}
\alias{print_pal}
\title{Plots a colour palette}
\usage{
print_pal(col_pal)
}
\arguments{
\item{col_pal}{a vector of colours}
}
\description{
Visually inspect the given colour palette.
}
\examples{
my_col_pal <- get_pal("caw_div_3")
print_pal(my_col_pal)
}
